Updated Cython section to use dask_mpi for exercise

+++ code/computePartMod.pyx
@@ -1,5 +1,5 @@
-cpdef int computePart(int size):
-  cdef int part=0
+cpdef long long computePart(int size):
+  cdef long long part=0
   cdef int i
   for i in range(size):
     part=part+i
